// Groundtruth-Simulator/config_generator/types/vehicleSettings.ts
import L from 'leaflet'

export type PositionTyp = {
  LatLng: L.LatLng
  Z?: number;
}


export type DestinationTyp = {
  position: PositionTyp;
  speed: number;
  error: number;
}


export type PointTyp = {
  position: PositionTyp
  lattitude: number;
  longtude: number;
  height?: number;
  
}


export type PropertiesTyp = {
  max_speed: number;
  max_force: number;
  max_altitude: number;
  position: PositionTyp;
}


export type VehicleTyp = {
  vehicle_id: number;
  vehicle_type: string;
  is_3D: boolean;
  action: string;
  properties: PropertiesTyp;
  destinations?: DestinationTyp[]
};